Migrate projects dropdown to resource dropdown
The new ResourceDropdownComponent is a superclass;
ProjectsDropdownComponent is now a subclass
that customizes a couple necessary values.
