+++ components/automate-ui/src/app/components/resource-dropdown/resource-dropdown.component.scss
@@ -1,0 +1,116 @@
+@import "~styles/variables";
+
+:host {
+  display: block;
+}
+
+// todo: update dropdown to have min-width of 250px and increase in width based on length of project name
+.dropdown-wrap {
+  position: relative;
+
+  #projects-selected {
+    // TODO (tc): the value for $chef-dark-grey here
+    // and what it says in the card description conflict.
+    // From the card:
+    //
+    // The value shown uses `chef-dark-gray` #212234
+    color: #212234;
+  }
+}
+
+.dropdown-button {
+  position: relative;
+  z-index: 2;
+  width: 100%;
+  padding: 18px;
+  background: $chef-white;
+  border: 1px solid $chef-light-grey;
+  border-radius: $global-radius;
+  text-align: left;
+  font-size: 14px;
+  color: $chef-dark-grey;
+  cursor: pointer;
+
+  &.active {
+    z-index: 4;
+    border-bottom-left-radius: 0;
+    border-bottom-right-radius: 0;
+  }
+
+  &[disabled] {
+    background: $chef-light-grey;
+    color: rgba($chef-primary-dark, 0.5);
+    cursor: default;
+    pointer-events: none;
+    opacity: 0.5;
+  }
+
+  &:focus {
+    outline: none;
+    border-color: $chef-grey;
+  }
+
+  > chef-icon {
+    position: absolute;
+    right: 18px;
+    top: 50%;
+    margin-top: -10px;
+    font-size: 23px;
+    color: $chef-primary-dark;
+  }
+}
+
+chef-dropdown {
+  width: 100%;
+  font-size: 14px;
+  z-index: 3;
+  border: 1px solid $chef-grey;
+  border-top: none;
+  border-top-left-radius: 0;
+  border-top-right-radius: 0;
+
+  #filter-container {
+    padding: 14px 18px;
+    border-bottom: 1px solid $chef-grey;
+  }
+
+  #project-container {
+    max-height: 210px;
+    overflow-y: scroll;
+  }
+
+  .no-projects {
+    padding: 14px 18px 0px;
+    font-style: italic;
+  }
+}
+
+chef-checkbox {
+  display: flex;
+  padding: 9px 18px;
+  align-items: flex-start;
+  white-space: normal; // prevents text from going outside of the element's container
+
+  &:first-child {
+    padding-top: 18px;
+  }
+
+  &:last-child {
+    padding-bottom: 18px;
+  }
+}
+
+.error {
+  padding: 8px 10px;
+  font-size: 10px;
+  background-color: $chef-critical;
+  color: $white;
+  border-radius: 0 0 $global-radius $global-radius;
+  box-shadow: 0 0 0 1px $chef-critical;
+}
+
+.help-text {
+  margin-top: 6px;
+  font-size: 12px;
+  line-height: 1.6;
+}
