+++ components/automate-ui/src/app/components/resource-dropdown/resource-dropdown.component.spec.ts
@@ -1,4 +1,6 @@
+import { CUSTOM_ELEMENTS_SCHEMA, EventEmitter } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { FormsModule } from '@angular/forms';
 
 import { ResourceDropdownComponent } from './resource-dropdown.component';
 
@@ -8,7 +10,9 @@ describe('ResourceDropdownComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ResourceDropdownComponent ]
+      declarations: [ ResourceDropdownComponent ],
+      imports: [ FormsModule ],
+      schemas: [ CUSTOM_ELEMENTS_SCHEMA ]
     })
     .compileComponents();
   }));
@@ -16,10 +20,42 @@ describe('ResourceDropdownComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(ResourceDropdownComponent);
     component = fixture.componentInstance;
+    component.projectsUpdated = new EventEmitter();
     fixture.detectChanges();
   });
 
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  describe('dropdown', () => {
+    beforeEach(() => {
+      component.active = true;
+      component.filteredProjects = [
+        {
+          name: 'Project 1',
+          id: 'project-1',
+          checked: false
+        },
+        {
+          name: 'Project 2',
+          id: 'project-2',
+          checked: true
+        }
+      ];
+      fixture.detectChanges();
+    });
+
+    it('displays a list of checkbox options', () => {
+      const options = Array.from(fixture.nativeElement.querySelectorAll('chef-checkbox'));
+      expect(options.length).toEqual(2);
+      options.forEach((option: HTMLInputElement, index: number) => {
+        const { name, checked } = component.filteredProjects[index];
+        expect(option.textContent).toEqual(name);
+        expect(option.checked).toEqual(checked);
+      });
+    });
+
+  });
+
 });
